単発と連射のクラスを分けた
