+++ Assets/Nishiwaki/Script/WeaponRangedAuto.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f17c9c6ca9813e42a6c721bb6d48ee5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
