+++ Assets/Nishiwaki/Script/WeaponRangedSemi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6abe3ebd94b35748a9ea60356defc0f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
